pyproject.toml
[project]
name = "neuralnetgame"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Mikeymoo5", email = "moodids@outlook.com"},
]
dependencies = [
    "tensorflow-io-gcs-filesystem==0.30.0",
    "tensorflow==2.11",
    "gymnasium>=0.29.1",
    "tf-agents==0.15.0",
    "pygame>=2.1.0",
    "click>=8.1.7",
    "numpy==1.21.6",
    "tensorflow-probability==0.19.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
